python/requirements.txt
azure-storage-blob==12.0.b1
azure-storage-file==12.0.b1
